Remove scan to check if down-conversion is required
The current heuristics to detect whether a down-conversion is required
rely on an additional scan to check if there are messages with a newer
message format than the max version required by the fetch request. With
the changes proposed KIP-724, we expect messages on disk to have format
version 2 in the common case. Given that, the heuristics are no longer
useful and we will always down-convert if the fetch request is v3 or
lower.